// disc-tracking-software/lib/pb/api.ts
export interface SyncBatch {
    /**
     * @generated from protobuf field: repeated disctracking.Ping pings = 1
     */
    pings: Ping[];
}
// ...
class SyncBatch$Type extends MessageType<SyncBatch> {
    constructor() {
        super("disctracking.SyncBatch", [
            { no: 1, name: "pings", kind: "message", repeat: 2 /*RepeatType.UNPACKED*/, T: () => Ping }
        ]);
    }
    create(value?: PartialMessage<SyncBatch>): SyncBatch {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.pings = [];
        if (value !== undefined)
            reflectionMergePartial<SyncBatch>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: SyncBatch): SyncBatch {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* repeated disctracking.Ping pings */ 1:
                    message.pings.push(Ping.internalBinaryRead(reader, reader.uint32(), options));
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: SyncBatch, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* repeated disctracking.Ping pings = 1; */
        for (let i = 0; i < message.pings.length; i++)
            Ping.internalBinaryWrite(message.pings[i], writer.tag(1, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message disctracking.SyncBatch
 */
export const SyncBatch = new SyncBatch$Type();